Configure client middleware on the server too
The sidekiq server is also a client of itself because jobs can enqueue
other jobs.  If we want these kinds of jobs to get the metrics recorded
by the client middleware, we need to configure the client-on-the-server
as well as the client-not-on-the-server to include the middleware.  This
is a little bit confusing, so see [the sidekiq wiki entry on configuring
middleware][middleware-wiki-entry] for more detail.

[middleware-wiki-entry]: https://github.com/mperham/sidekiq/wiki/Middleware#client-middleware-registered-in-both-places